requirements.txt
matplotlib
pandas
pytorch-lightning
tensorboard
torch_tb_profiler
ipykernel
lpips 
opencv-python
plotly
scikit-image
nbformat
